requirements.txt
ml-collections==0.1.0
tensorflow-gan==2.0.0
tensorflow-probability==0.21.0
tensorflow_io
tensorflow_datasets==4.9.3
tensorflow==2.13.0
tensorflow-addons==0.23.0
tensorboard==2.13.0
absl-py>=1.4.0
torch>=1.7.0
torchvision
ninja
scipy>=1.9
